backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pymongo==4.7.2
python-dotenv==1.0.1
pydantic>=2.11.7,<3.0.0